gcp-cloud-functions/requirements.txt
google-cloud-build
google-cloud-pubsub==2.13.7
google-cloud-storage==2.5.0